Add user feedback support using Grazie platform
