+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/RequestManager.kt
@@ -13,14 +13,41 @@ abstract class RequestManager {
 
     open val log: Logger = Logger.getInstance(this.javaClass)
 
-    abstract fun request(
+    /**
+     * Sends a request to LLM with the given prompt and returns the generated TestSuite.
+     *
+     * @param prompt the prompt to send to LLM
+     * @param indicator the progress indicator to show progress during the request
+     * @param packageName the name of the package for the generated TestSuite
+     * @param project the project associated with the request
+     * @param llmErrorManager the error manager to handle errors during the request
+     * @param isUserFeedback indicates if this request is a test generation request or a user feedback
+     * @return the generated TestSuite, or null and prompt message
+     */
+    open fun request(
         prompt: String,
         indicator: ProgressIndicator,
         packageName: String,
         project: Project,
         llmErrorManager: LLMErrorManager,
         isUserFeedback: Boolean = false
-    ): Pair<String, TestSuiteGeneratedByLLM?>
+    ): Pair<String, TestSuiteGeneratedByLLM?>{
+        // save the prompt in chat history
+        chatHistory.add(ChatMessage("user", prompt))
+
+        // Send Request to LLM
+        log.info("Sending Request ...")
+        val testsAssembler = send(prompt, indicator,project,llmErrorManager)
+
+        // we remove the user request because we dont users requests in chat history
+        if (isUserFeedback)
+            chatHistory.removeLast()
+
+        return when(isUserFeedback){
+            true -> processUserFeedbackResponse(testsAssembler, packageName)
+            false -> processResponse(testsAssembler, packageName)
+        }
+    }
 
     open fun processResponse(
         testsAssembler: TestsAssembler,
